Tests: the four Kotlin warnings in the test sources are gone, so the CI compile can hold them to zero too
The CI warning gate never compiled the test sources, and four warnings lived
there (the Unit tests step printed them, and nothing counted them):
- EmbedContainerTest.kt:162  `Object() as T` — "This class is not recommended
  for use in Kotlin. Use 'kotlin.Any' instead." → `Any() as T`.
- EmbedContainerTest.kt:298  `Automatic is EmbedIsolatedNavigation` — "Check
  for instance is always 'true'." The case being one of the sealed type's is
  a compile-time fact: it is now a typed assignment (compiles only if so),
  plus a runtime check that it is not the Custom case.
- DataSectionDefaultsTest.kt:57, :73  `resolveNumber(…)!!.toDouble()` —
  "Redundant call of conversion method." resolveNumber returns Double?.

With every Kotlin warning an error (the next commit's CI init script), main
and test sources compile clean: Gradle 9.4.1 / Kotlin 2.4.10, the five
compile tasks rc 0; library 223 and library-dynamic 452 unit tests, 0
failures. Putting one of these back fails that compile, naming its line.

+++ library-dynamic/src/test/kotlin/com/kotlinjsonui/dynamic/DataSectionDefaultsTest.kt
@@ -54,7 +54,7 @@ class DataSectionDefaultsTest {
         assertEquals("Osaka", DataBindingContext.resolveString("@{profile.address.city}", effective))
         assertEquals(
             34.7,
-            DataBindingContext.resolveNumber("@{profile.address.geo.lat}", effective)!!.toDouble(),
+            DataBindingContext.resolveNumber("@{profile.address.geo.lat}", effective)!!,
             1e-9
         )
     }
@@ -70,7 +70,7 @@ class DataSectionDefaultsTest {
         assertEquals("green", DataBindingContext.resolveString("@{tags[1]}", effective))
         assertEquals(
             20.0,
-            DataBindingContext.resolveNumber("@{profile.scores[1]}", effective)!!.toDouble(),
+            DataBindingContext.resolveNumber("@{profile.scores[1]}", effective)!!,
             1e-9
         )
         // Out-of-range index stays unresolved (no crash, no fallback value)
